move slack operations to shared
